MSRZG3S: add board config for the RZ/G3S OSM from ARIES Embedded
The MSRZG3S OSMs are available as MSRZG3S-BAA with 1 GB of LPDDR4
and eMMC and MSRZG3S-A0A with just 512 MB and without eMMC. They
can be mounted on the OSMEVK or the G3Sberry board.

The support of derived from the Renesas "smarc-rzg3s" board, which
is very similar.

Signed-off-by: Wolfgang Grandegger <wg@aries-embedded.de>